feat(M3): Introduce checkbox API; refactor(SimpleIME): Migrate all checkbox API to M3 style checkbox;

+++ common/imguiex/ErrorNotifier.cpp
@@ -80,7 +80,7 @@ void ErrorNotifier::Show()
             if (message.level >= m_currentLevel)
             {
                 renderMessage(message, static_cast<size_t>(i));
-                if (static_cast<size_t>((current - message.timestamp).count()) >= (m_duration.count()))
+                if (static_cast<size_t>((current - message.timestamp).count()) >= static_cast<size_t>((m_duration.count())))
                 {
                     message.confirmed = true;
                 }
